+++ NEWS.md
@@ -1,6 +1,8 @@
 # Version 1.3-2 (development)
 
-* add `SDM.RStests` implementation of Rao's score tests for spatial Durbin models (Koley and Bera, 2024) 
+* change `lm.LMtests` to `lm.RStests` and re-name Lagrange multiplier to Rao's score
+
+* add `SD.RStests` implementation of Rao's score tests for spatial Durbin models (Koley and Bera, 2024) and for SDEM models (Koley, forthcoming)
 
 * add `style` to `sn2listw` use in `tri2nb`