+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * add `SD.RStests` implementation of Rao's score tests for spatial Durbin models (Koley and Bera, 2024) and for SDEM models (Koley, forthcoming)
  
+* #143 `row.names` pass-through in `poly2nb` corrected, harmonised `row.names` pass-through also in `nbdists` and `dnearneigh`
+
 * #139 add `na.action` argument to `geary.test`, `geary.mc` and `globalG.test`
 
 * add `style` to `sn2listw` use in `tri2nb`