+++ NEWS.md
@@ -1,4 +1,4 @@
-# Version 1.3-2 (development)
+# Version 1.3-3 (development)
 
 * change `lm.LMtests` to `lm.RStests` and re-name Lagrange multiplier to Rao's score; add `GNM_` prefix to test names if the input object inherits from `SlX` created by `spatialreg::lmSLX` (Koley, forthcoming)